+++ tests/001_functionals/suite_005_controlbase.py
@@ -4,6 +4,7 @@
 
 import pytest
 from qacode.core.exceptions.control_exception import ControlException
+from qacode.core.exceptions.core_exception import CoreException
 from qacode.core.testing.test_info import TestInfoBotUnique
 from qacode.core.webs.controls.control_base import ControlBase
 from qautils.files import settings
@@ -136,18 +137,18 @@ class TestControlBase(TestInfoBotUnique):
             "on_instance_search": None,
             "auto_reload": None,
         }
-        with pytest.raises(ControlException):
+        with pytest.raises(CoreException):
             ControlBase(self.bot, **cfg)
 
     @pytest.mark.skipIf(SKIP_CONTROLS, SKIP_CONTROLS_MSG)
     def test_instance_raises_nonebot(self):
         """Testcase: test_instance_raises_nonebot"""
-        self.assert_raises(ControlException, ControlBase, None)
+        self.assert_raises(CoreException, ControlBase, None)
 
     @pytest.mark.skipIf(SKIP_CONTROLS, SKIP_CONTROLS_MSG)
     def test_instance_raises_nonesettings(self):
         """Testcase: test_instance_raises_nonesettings"""
-        self.assert_raises(ControlException, ControlBase, self.bot)
+        self.assert_raises(CoreException, ControlBase, self.bot)
 
     @pytest.mark.skipIf(SKIP_CONTROLS, SKIP_CONTROLS_MSG)
     def test_property_gettext(self):